+++ package.json
@@ -33,6 +33,7 @@
     "component-each": "^0.2.6",
     "crypto-js": "^4.0.0",
     "dotenv": "^8.2.0",
+    "generic-pool": "^3.7.1",
     "get-value": "^3.0.1",
     "handlebars": "^4.7.6",
     "is": "^3.1.0",
